Rename documented classes, polish header javadoc
