+++ reactor-core-micrometer/src/main/java/reactor/core/observability/micrometer/DocumentedMeterListenerMeters.java
@@ -18,12 +18,17 @@ package reactor.core.observability.micrometer;
 
 import io.micrometer.common.docs.KeyName;
 import io.micrometer.core.instrument.Meter;
+import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.docs.DocumentedMeter;
 
 /**
- * Meters and tags used by {@link MicrometerMeterListener}.
+ * Meters and tags used by {@link Micrometer#metrics(MeterRegistry)}.
  */
-public enum PublisherMeters implements DocumentedMeter {
+/*
+NOTE that javadocs avoid using links / code taglets as these are replaced with blanks when
+the docs are generated by Micrometer's tool
+*/
+public enum DocumentedMeterListenerMeters implements DocumentedMeter {
 
 	/**
 	 * Counts the number of events received from a malformed source (ie an onNext after an onComplete).
