+++ reactor-core-micrometer/src/main/java/reactor/core/observability/micrometer/DocumentedObservationListenerTags.java
@@ -17,13 +17,18 @@
 package reactor.core.observability.micrometer;
 
 import io.micrometer.common.docs.KeyName;
+import io.micrometer.observation.ObservationRegistry;
 import io.micrometer.observation.docs.DocumentedObservation;
 
 /**
- * Documentation of {@link MicrometerObservationListener} tags and of the anonymous variant
+ * Documentation of {@link Micrometer#observation(ObservationRegistry)} tags and of the anonymous variant
  * of the observation (no {@link reactor.core.publisher.Flux#name(String)}).
  */
-public enum PublisherObservations implements DocumentedObservation {
+/*
+NOTE that javadocs avoid using links / code taglets as these are replaced with blanks when
+the docs are generated by Micrometer's tool
+*/
+public enum DocumentedObservationListenerTags implements DocumentedObservation {
 
 	/**
 	 * Anonymous version of the Micrometer.observation(), when the sequence hasn't been
@@ -74,20 +79,20 @@ public enum PublisherObservations implements DocumentedObservation {
 		/**
 		 * {@link #STATUS} for when the subscription to the sequence was cancelled.
 		 */
-		public static final String TAG_STATUS_CANCELLED = PublisherMeters.TerminationTags.TAG_STATUS_CANCELLED;
+		public static final String TAG_STATUS_CANCELLED = DocumentedMeterListenerMeters.TerminationTags.TAG_STATUS_CANCELLED;
 		/**
 		 * {@link #STATUS} for when the sequence completed with values.
 		 */
-		public static final String TAG_STATUS_COMPLETED = PublisherMeters.TerminationTags.TAG_STATUS_COMPLETED;
+		public static final String TAG_STATUS_COMPLETED = DocumentedMeterListenerMeters.TerminationTags.TAG_STATUS_COMPLETED;
 		/**
 		 * {@link #STATUS} for when the sequence completed without value (no onNext).
 		 */
-		public static final String TAG_STATUS_COMPLETED_EMPTY = PublisherMeters.TerminationTags.TAG_STATUS_COMPLETED_EMPTY;
+		public static final String TAG_STATUS_COMPLETED_EMPTY = DocumentedMeterListenerMeters.TerminationTags.TAG_STATUS_COMPLETED_EMPTY;
 		/**
 		 * {@link #STATUS} for when the sequence terminated with an error. The {@link io.micrometer.observation.Observation#error(Throwable)}
 		 * method is used to capture the exception.
 		 */
-		public static final String TAG_STATUS_ERROR = PublisherMeters.TerminationTags.TAG_STATUS_ERROR;
+		public static final String TAG_STATUS_ERROR = DocumentedMeterListenerMeters.TerminationTags.TAG_STATUS_ERROR;
 	}
