+++ reactor-core-micrometer/src/main/java/reactor/core/observability/micrometer/DocumentedTimedSchedulerMeters.java
@@ -18,12 +18,19 @@ package reactor.core.observability.micrometer;
 
 import io.micrometer.common.docs.KeyName;
 import io.micrometer.core.instrument.Meter;
+import io.micrometer.core.instrument.MeterRegistry;
 import io.micrometer.core.instrument.docs.DocumentedMeter;
 
+import reactor.core.scheduler.Scheduler;
+
 /**
- * Meters and tags used by {@link TimedScheduler}.
+ * Meters and tags used by {@link Micrometer#timedScheduler(Scheduler, MeterRegistry, String)}.
  */
-public enum SchedulerMeters implements DocumentedMeter {
+/*
+NOTE that javadocs avoid using links / code taglets as these are replaced with blanks when
+the docs are generated by Micrometer's tool
+*/
+public enum DocumentedTimedSchedulerMeters implements DocumentedMeter {
 
 	/**
 	 * Counter that increments by one each time a task is submitted (via any of the
